chore: move clubs.d.ts
