+++ components/custom/club-card.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import type { PropType } from 'vue'
-import type { Club } from '@/content/clubs'
+import type { Club } from '~/types/clubs'
 import { Button } from '@/components/ui/button'
 import { Card, CardContent, CardDescription, CardFooter, CardHeader, CardTitle } from '@/components/ui/card'
 import { cleanHTML, cn } from '@/lib/utils'
